thomas/software_defined_radio/requirements.txt
aiofiles==24.1.0
bitarray==3.0.0
bitstring==4.3.0
contourpy==1.3.1
cycler==0.12.1
filelock==3.16.1
fonttools==4.55.2
fsspec==2024.12.0
iniconfig==2.0.0
Jinja2==3.1.5
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.3
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pluggy==1.5.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.2.0
pyperclip==1.9.0
PyRect==0.2.0
pyrtlsdr==0.3.0
PyScreeze==1.0.1
pytest==8.3.4
pytest-qt==4.4.0
python-dateutil==2.9.0.post0
python3-xlib==0.15
pytweening==1.2.0
pytz==2025.1
QDarkStyle==3.2.3
QtPy==2.4.2
scapy==2.6.1
six==1.17.0
sympy==1.13.1
torch==2.5.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2025.1
